+++ src/tasks/client-integrity-task.cc
@@ -24,6 +24,7 @@
 #include <vector>
 
 #include "clients/client.h"
+#include "core/configuration.h"
 #include "core/registry.h"
 #include "logging/log.h"
 #include "tokenization/token.h"
@@ -36,18 +37,59 @@ ClientIntegrityTask::ClientIntegrityTask(Registry* registry,
 {
 }
 
-void ClientIntegrityTask::performCleanup()
+void ClientIntegrityTask::performCleanup(const std::string& clientId)
+{
+    auto iter = m_registry->clients().find(clientId);
+    if (iter == m_registry->clients().end()) {
+        // This is possible in case of "unknown"
+        return;
+    }
+    Client* client = &(iter->second);
+    if (!client->isAlive()) {
+        // Do not use m_registry.removeClient, instead,
+        // manually remove as we need to manually increment
+        // iterator as erase invalidates it
+        std::lock_guard<std::recursive_mutex> lock(m_registry->mutex());
+        m_registry->clients().erase(iter);
+    } else {
+        if (!client->backupKey().empty()) {
+            RVLOG(RV_WARNING) << "Removing backup key for [" << client->id() << "]";
+            client->setBackupKey("");
+        }
+    }
+}
+
+void ClientIntegrityTask::execute()
 {
     auto* list = &(m_registry->clients());
     for (auto clientIter = list->begin(); clientIter != list->end();) {
         Client* client = &(clientIter->second);
         if (!client->isAlive()) {
-            RLOG(INFO) << "Client expired " << client->id();
-            // Do not use m_registry.removeClient, instead,
-            // manually remove as we need to manually increment
-            // iterator as erase invalidates it
-            std::lock_guard<std::recursive_mutex> lock(m_registry->mutex());
-            m_registry->clients().erase(clientIter++);
+
+            std::lock_guard<std::mutex> pausedClientsLock(m_mutex);
+            auto pos = m_pausedClients.find(client->id());
+            if (pos == m_pausedClients.end()) {
+
+                // here we check for unknown clients
+                // we may be removing a "paused" unknown client
+                // as m_pausedClients for unknown client will be "unknown"
+                if (!client->isKnown() && m_pausedClients.find(Configuration::UNKNOWN_CLIENT_ID) != m_pausedClients.end()) {
+                    // yes we are surely removing a paused unknown client
+                    // that we shouldn't do
+                    RLOG(INFO) << "Unknown client [" << client->id() << "] expired - Paused removal";
+                    ++clientIter; // skip and move on
+                } else {
+                    RLOG(INFO) << "Client [" << client->id() << "] expired";
+                    // Do not use m_registry.removeClient, instead,
+                    // manually remove as we need to manually increment
+                    // iterator as erase invalidates it
+                    std::lock_guard<std::recursive_mutex> lock(m_registry->mutex());
+                    m_registry->clients().erase(clientIter++);
+                }
+            } else {
+                RLOG(INFO) << "Client [" << client->id() << "] expired - Paused removal";
+                ++clientIter; // skip and move on
+            }
         } else {
             if (!client->backupKey().empty()) {
                 RVLOG(RV_WARNING) << "Removing backup key for [" << client->id() << "]";
@@ -55,16 +97,5 @@ void ClientIntegrityTask::performCleanup()
             }
             ++clientIter;
         }
-    }
-}
-
-void ClientIntegrityTask::execute()
-{
-    if (m_performCleanUpOnSchedule) {
-        performCleanup();
-    } else {
-        // This only marks last execution
-        // it's processed in log request handler to prevent any damange
-        RLOG(DEBUG) << "Paused for scheduled clean up";
     }
 }