+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Updates
 - New format specifier `%session_id`
+- Client integrity tasks are paused for specific clients being processed instead of all clients
 
 ## [1.6.0] - 27-02-2018
 ### Updates